+++ bank/src/main/java/com/fisa/bank/common/config/security/authorization/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.config.security.auth;
+package com.fisa.bank.common.config.security.authorization;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
