+++ bank/src/main/java/com/fisa/bank/common/config/security/authorization/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.config.security.auth;
+package com.fisa.bank.common.config.security.authorization;
 
 import lombok.RequiredArgsConstructor;
